rename login routes and default strategy file
