+++ demo/server/views/login.hbs
@@ -27,7 +27,7 @@
 
         <div id="form" >
           <div id="form-wrapper">
-            <form id="form-login" action="/login" method="post">
+            <form id="form-login" action="/cookie-login" method="post">
               <div>
                 <div>
                   <label for="username">Username or email</label>
